First policy: forbid JAL instructions
As a first instance of a marginally more interesting [mpstep], perform
some simple checks on the machine state, namely the currently
executing instruction, and disallow some of these statically. This
requires us to replicate some instruction decoding work, which can
hopefully be done more compactly.

The natural continuation of this idea is to allow only a subset of
call sites declared in the policy state. At the moment, only the
definition of the policy type has been modified to accept a list of
call sites, but this information is not yet used.