Eager policy
Change base policy to allow only reads and writes in the current stack
frame, assuming unique activation ids.

This requires the blessed header sequence to tag the stack frames as
it creates them. At the moment, the policy performs this operation as
the SP is incremented, even though a PIPE machine would be unable to
do so.