Tag support for global variables
It is easy to consider a simple model of observations where certain
designated positions in memory act as observable ports. These are
outside the stack and can be at least written (and possibly read) by
everyone. Special rules to allow loads and stores to these
locations (possibly with further restrictions, as with other memory
operations, e.g., those that attempt SP-based misbehavior) are quickly
added.