Add changed address to visible event
Without this addition, it is difficult to monitor more than one visible
"port" in memory. It is clear that the extensive lazy property that
monitors the entirety of the data memory will essentially always fail
after a call: the variant machine will have varied those parts of the
stack that correspond to the newly created stack frame; when the blessed
sequence attempts to initialize them, a tagging assignment that does not
modify the value in the original machine will reset the modified value
in the variant, leading to diverging events in an observed position.