Consistent use of eager tags
In the store rule, split the treatment of untagged and tagged memory for
regular stores, and require that in the latter case the depth/activation
tags of PC and memory address coincide.

Because stores are tagging memory with Tpc, it makes no sense to look
for Tstack during these checks; this tag confusion is of no consequence,
but should be revisited later.