Restore and adapt stack and code maps
Now we can go all the way back to the machine model and reintroduce
stack maps (as they were before) and code maps (using the new type of
operations instead of the old code annotations). These will be needed
to support various aspects of testing.

The use of lists of operations makes the detection of call and return
points a little more interesting, in that the types do not carry the
information as unambiguously, and the encoding does not automatically
preclude nonsensical operation lists.