Prototype of more efficient tester
Existing properties become inefficient very quickly, partly because,
e.g., new instances of the some properties and checks are added at call
points, and these step the machine independently from all others.

Instead, we consider a program execution framework that maintains a
stack-like list of tests, is in charge of performing a single canonical
run of the machine and performs all checks at the appropriate levels.