Testing failures
Using [genMach] to test confidentiality now allows us to find programs
that fail the property; at least some of these are surprising. For
example, even an empty function can trigger such a violation. Some of
the programs, e.g., the counterexample included here, jump to the
function while missing its call point.

It could be useful to run the property on the fixed lazy policy and
check the violations that remain.