Tweak to original-lazy policy and confidentiality
Restoration of the next ID on return rule of the original policy does
not seem to take into account the fact that ID generation is based on
the successor of the so-called next ID field, therefore correct by
using the predecessor of the (successor of the) next ID at restore time.

Paired with the proposed fix to the property, the second and third examples
appear to work as expected.