Relative arguments
See comments for general design notes.

Generally speaking, stack memory tags are decorated with the kind of
data they contain. Currently, this can be regular stack data or
arguments passed to a function's callees.

Code tags are extended with a new category reserved for instructions
used to store the arguments of an upcoming function call in the caller's
frame. We assume that e.g. the compiler correctly annotates code to
denote these argument-setting instructions.

Regular policy rules for loads and stores are duplicated. The original
cases cover "regular" stack data (i.e., not relative arguments, or other
types as those are added). The new sub-cases cover instructions that
manipulation of relative arguments passed on the stack. In particular,
store instructions tagged with the new [Tsetarg] tag set the kind of
stack data accordingly. Conversely, load instructions on relative
arguments belonging to the parent frame are allowed to proceed.

Generation currently supports relative arguments only. On the one hand,
stack-based loads may opt to read one of the current function's stack
arguments, thereby accessing the caller's frame. On the other, when a
caller makes a call to a function that takes arguments on the stack,
the call instruction is prefixed by a sequence of specially tagged
stores that mark those data as arguments.

Several generalizations are possible, including on the number of
relative arguments (note that this can increase the size of the
generated functions quite a bit, albeit in predictable and mostly
uninteresting ways), allocation of stack frames (all in one go at call
time or in several steps, e.g., before each sub-call), invalidation of
argument data (via regular, unprivileged stores to the same locations),
etc.