Checkpoint on callee-saved testing
The one actual hack used here is the addition of an extra word at the
bottom of each frame, for homogeneity, left generally as empty padding.
The reason for this irregularity is that the main function uses the
bottom word of its stack frame to save a "return address of origin" and
does not follow the convention that the return address is stored in the
top word of the caller's stack frame (morally at least, as alloc goes).